refactor: move `typeorm-query` to its folder
