+++ src/typeorm/filter/field/field-filter.ts
@@ -3,7 +3,7 @@ import { Not } from 'typeorm'
 import { AbstractFilter } from '../filter'
 import { LookupFilter } from './lookup.enum'
 import { ExpressQuery } from '../../../express-query'
-import { TypeORMQuery } from '../../../typeorm-query'
+import { TypeORMQuery } from '../../query'
 
 interface FilterConfig {
   query: ExpressQuery
